Day9/requirements.txt
streamlit
pandas
chromadb
langchain
langchain-community
langchain-google-genai
langchain-chroma
PyPDF2
python-docx
google-generativeai
sentence-transformers